pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "automaxo"
version = "0.1.0"
description = "Using LLM to help in curation of Medical Action Ontology (MaXO)"
authors = [
    "Enock Niyonkuru <enock.niyonkuru@trincoll.edu>", 
    "Peter Robinson <peter.robinson@jax.org>"
    ]
license = "MIT"
readme = "README.md"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
]
keywords = ["LLM", "NLP", "MaXO"]

[tool.poetry.dependencies]
python = "^3.9"
biopython = "1.79"
sparqlwrapper = "^2.0.0"
ontogpt = "0.3.8"

[tool.poetry.dev-dependencies]
# Add development dependencies here

[tool.poetry.scripts]
# Define any entry points for command-line scripts here
